+++ lib/keycloak-admin/client/client_role_mappings_client.rb
@@ -13,10 +13,10 @@ module KeycloakAdmin
       JSON.parse(response).map { |role_as_hash| RoleRepresentation.from_hash(role_as_hash) }
     end
 
-    def save(role_representations)
+    def save(role_representation_list)
       execute_http do
         RestClient::Resource.new(base_url, @configuration.rest_client_options).post(
-          role_representations.to_json, headers
+          role_representation_list.to_json, headers
         )
       end
     end